pep(python)/Portfolio_project/api_app/serializers.py
from rest_framework import serializers
from portfolio.models import Project, ContactMessage

class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = ['id', 'title', 'slug', 'description', 'tech_stack', 'github', 'live_link', 'image', 'created_at']

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactMessage
        fields = ['id', 'name', 'email', 'message', 'timestamp']