pep(python)/Portfolio_project/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-environ==0.13.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
